backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
aiohttp==3.8.4
pydantic==1.10.12
python-dotenv==1.0.0
